Refactor all chat database methods into BMInterfaceGameChat module
