+++ src/api/ApiResponder.php
@@ -422,7 +422,7 @@ class ApiResponder {
         if (!(array_key_exists('edit', $args))) {
             $args['edit'] = FALSE;
         }
-        $retval = $interface->game()->submit_chat(
+        $retval = $interface->game_chat()->submit_chat(
             $_SESSION['user_id'],
             $args['game'],
             $args['edit'],
