+++ test/src/engine/BMInterfaceGameChatTest.php
@@ -1,0 +1,14 @@
+<?php
+
+require_once 'BMInterfaceTestAbstract.php';
+
+class BMInterfaceGameChatTest extends BMInterfaceTestAbstract {
+
+    protected function init() {
+        $this->object = new BMInterfaceGameChat(TRUE);
+    }
+
+    public function test_dummy() {
+        // currently, there are no interface-level tests of game action functionality
+    }
+}
